updated todos controller

+++ controllers/todos.js
@@ -1,10 +1,11 @@
 const db = require('../database')
-const getToken = require('../middleware/getToken')
-
+const getCurrUser = require('../middleware/getCurrUser')
+const verifyToken = require('../middleware/vetifyToken')
 module.exports = {
     getTodos: async (req, res) => {
         try {
-            const email = getToken(req.headers.authorization)
+            const email = getCurrUser(req.headers.authorization)
+            if(!email) return res.status(404).send({msg: 'User not found'})
             const result = await db.query('SELECT * FROM TODOS WHERE email = ?', [email])
             res.status(200).send(result[0])
             
@@ -14,11 +15,14 @@ module.exports = {
         }
     },
     createTodo: async (req, res) => {
-        const { title, due_date, description } = req.body
-        const email = getToken(req.headers.authorization)
+        const { id, title, due_date, description } = req.body
+        const token = req.headers.authorization
+        const response = verifyToken(token)
+        if(!response) return res.status(401).send({ msg: 'Token not valid or not found, plz login again'})
+        const email = getCurrUser(req.headers.authorization)
         if(title){
             try {
-                await db.query('INSERT INTO TODOS (title, due_date, description, email) VALUES (?, ?, ?, ?)', [title, due_date, description, email])
+                await db.query('INSERT INTO TODOS (id, title, due_date, description, email) VALUES (?, ?, ?, ?, ?)', [id, title, due_date, description, email])
                 res.status(201).send({msg: 'TODO CREATED SUCCESSFULLY'})  
             } catch (error) {
                 console.log(error)
@@ -28,9 +32,11 @@ module.exports = {
     updateTodo: async (req, res) => {
         try {
             const data = req.body
+            const token = req.headers.authorization
+            const response = verifyToken(token)
+            if(!response) return res.status(401).send({ msg: 'Token not valid or not found, plz login again'})
             const id = req.params.id
             await db.query('UPDATE todos SET ?  WHERE id= ?',[data, id]);
-            //await db.query('UPDATE todos SET title=?, due_date=?, description=?, completed=?, position=? WHERE id= ?',[title, due_date, description,completed, position, id]);
             res.status(200).send({msg: 'Todo updated successfully'})
 
         } catch (error) {
@@ -41,6 +47,9 @@ module.exports = {
     },
     deleteTodo: async (req, res) => {
         try {
+            const token = req.headers.authorization
+            const response = verifyToken(token)
+            if(!response) return res.status(401).send({ msg: 'Token not valid or not found, plz login again'})
             await db.query('DELETE FROM todos WHERE id = ?', [req.params.id])
             res.status(202).send({msg: 'TODO DELETED SUCCESSFULLY'})
         } catch (error) {
